TECH-4819: Move fibered_mysql2_adapter.rb into lib/active_record/connection_adapters/
