+++ lib/active_record/connection_adapters/fibered_mysql2_adapter.rb
