+++ spec/unit/fibered_mysql2_adapter_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../lib/fibered_mysql2/fibered_mysql2_adapter'
+require_relative '../../lib/active_record/connection_adapters/fibered_mysql2_adapter'
 
 RSpec.describe FiberedMysql2::FiberedMysql2Adapter do
   let(:client) { Mysql2::EM::Client.new }
